feat: FormProvider로 감싸서 전역으로 관리, AuctionInputs 타입 분리
