+++ src/app/create/basicauction/types/InputTypes.ts
@@ -1,0 +1,13 @@
+type ImageUpload = {
+  file: File;
+};
+
+export type AuctionInputs = {
+  nameRequired: string;
+  startPriceRequired: string;
+  bidUnitRequired: string;
+  imagesRequired: ImageUpload[];
+  contentRequired: string;
+  startDateRequired: string;
+  endDateRequired: string;
+};
